LayoutTests/http/tests/media/media-source/video-media-source-add-and-remove-buffers-expected.txt

Tests webkitSourceAddId() & webkitSourceRemoveId() methods

Test adding an ID while closed.
Got expected exception Error: InvalidStateError: DOM Exception 11

running testAddBufferFailureCases
EVENT(webkitsourceopen)
Test empty type.
Got expected exception Error: InvalidAccessError: DOM Exception 15
Test an unsupported type.
Got expected exception Error: NotSupportedError: DOM Exception 9
Test a supported type with an unsupported codec.
Got expected exception Error: NotSupportedError: DOM Exception 9
Test reaching sourceID limit.
Test that SourceBuffers can't be added while in the ended state.
Got expected exception Error: InvalidStateError: DOM Exception 11

running testRemoveNullBuffer
EVENT(webkitsourceopen)
Test null buffer case
Got expected exception Error: InvalidAccessError: DOM Exception 15

running testRemoveAgain
EVENT(webkitsourceopen)
Test removing a buffer that was already removed.
Got expected exception Error: InvalidStateError: DOM Exception 11

running testRemoveBufferAfterEnded
EVENT(webkitsourceopen)
Test that a buffer can be removed while in the ended state.

running testAddBufferAfterRemoving
EVENT(webkitsourceopen)
Test that a buffer can be added again after it is removed.

running testAppendFailureCases
EVENT(webkitsourceopen)
Test a successful append.
Test append with a null buffer.
Got expected exception Error: InvalidAccessError: DOM Exception 15
Test append after buffer has been removed.
Got expected exception Error: InvalidStateError: DOM Exception 11
END OF TEST
